pyproject.toml
[project]
name = "llm-recsys"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1",
    "faiss-cpu>=1.7",
    "numpy>=1.24",
    "pandas>=2.0",
    "lightgbm>=4.0",
    "xgboost>=2.0",
    "sentence-transformers>=2.2",
    "anthropic>=0.25",
    "python-dotenv>=1.0",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4", "isort>=5.13"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["evals"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.isort]
profile = "black"
line_length = 100
